added error handling controller and changed to jpa repository for user and email repository.

+++ src/main/java/com/lms/lmsapi/service/EmailService.java
@@ -7,8 +7,8 @@ import com.lms.lmsapi.entity.Email;
 public interface EmailService {
 
     List<Email> getEmails();
-    Email getEmailByUserId(int id);
+    Email getEmailByUserId(Long id);
     Email createUserEmail(Email email);
     Email updateEmailByUserId(Email email);
-    void deleteUserEmail(int userId);    
+    void deleteUserEmail(Long userId);    
 }
